RavenDB-5615 moved indexes/start and stop to /admin
