+++ src/Raven.Server/Documents/Handlers/Admin/AdminIndexHandler.cs
@@ -1,4 +1,5 @@
-﻿using System.Net;
+﻿using System;
+using System.Net;
 using System.Threading.Tasks;
 using Raven.Server.Json;
 using Raven.Server.Routing;
@@ -37,6 +38,100 @@ namespace Raven.Server.Documents.Handlers.Admin
                 writer.WriteOperationId(context, operationId);
             }
 
+            return Task.CompletedTask;
+        }
+
+        [RavenAction("/databases/*/admin/indexes/stop", "POST")]
+        public Task Stop()
+        {
+            var types = HttpContext.Request.Query["type"];
+            var names = HttpContext.Request.Query["name"];
+            if (types.Count == 0 && names.Count == 0)
+            {
+                HttpContext.Response.StatusCode = (int)HttpStatusCode.NoContent;
+                Database.IndexStore.StopIndexing();
+                return Task.CompletedTask;
+            }
+
+            if (types.Count != 0 && names.Count != 0)
+                throw new ArgumentException("Query string value 'type' and 'names' are mutually exclusive.");
+
+            if (types.Count != 0)
+            {
+                if (types.Count != 1)
+                    throw new ArgumentException("Query string value 'type' must appear exactly once");
+                if (string.IsNullOrWhiteSpace(types[0]))
+                    throw new ArgumentException("Query string value 'type' must have a non empty value");
+
+                if (string.Equals(types[0], "map", StringComparison.OrdinalIgnoreCase))
+                {
+                    Database.IndexStore.StopMapIndexes();
+                }
+                else if (string.Equals(types[0], "map-reduce", StringComparison.OrdinalIgnoreCase))
+                {
+                    Database.IndexStore.StopMapReduceIndexes();
+                }
+                else
+                {
+                    throw new ArgumentException("Query string value 'type' can only be 'map' or 'map-reduce' but was " + types[0]);
+                }
+            }
+            else if (names.Count != 0)
+            {
+                if (names.Count != 1)
+                    throw new ArgumentException("Query string value 'name' must appear exactly once");
+                if (string.IsNullOrWhiteSpace(names[0]))
+                    throw new ArgumentException("Query string value 'name' must have a non empty value");
+
+                Database.IndexStore.StopIndex(names[0]);
+            }
+
+            HttpContext.Response.StatusCode = (int)HttpStatusCode.NoContent;
+            return Task.CompletedTask;
+        }
+
+        [RavenAction("/databases/*/admin/indexes/start", "POST")]
+        public Task Start()
+        {
+            var types = HttpContext.Request.Query["type"];
+            var names = HttpContext.Request.Query["name"];
+            if (types.Count == 0 && names.Count == 0)
+            {
+                HttpContext.Response.StatusCode = (int)HttpStatusCode.NoContent;
+                Database.IndexStore.StartIndexing();
+                return Task.CompletedTask;
+            }
+
+            if (types.Count != 0 && names.Count != 0)
+                throw new ArgumentException("Query string value 'type' and 'names' are mutually exclusive.");
+
+            if (types.Count != 0)
+            {
+                if (types.Count != 1)
+                    throw new ArgumentException("Query string value 'type' must appear exactly once");
+                if (string.IsNullOrWhiteSpace(types[0]))
+                    throw new ArgumentException("Query string value 'type' must have a non empty value");
+
+                if (string.Equals(types[0], "map", StringComparison.OrdinalIgnoreCase))
+                {
+                    Database.IndexStore.StartMapIndexes();
+                }
+                else if (string.Equals(types[0], "map-reduce", StringComparison.OrdinalIgnoreCase))
+                {
+                    Database.IndexStore.StartMapReduceIndexes();
+                }
+            }
+            else if (names.Count != 0)
+            {
+                if (names.Count != 1)
+                    throw new ArgumentException("Query string value 'name' must appear exactly once");
+                if (string.IsNullOrWhiteSpace(names[0]))
+                    throw new ArgumentException("Query string value 'name' must have a non empty value");
+
+                Database.IndexStore.StartIndex(names[0]);
+            }
+
+            HttpContext.Response.StatusCode = (int)HttpStatusCode.NoContent;
             return Task.CompletedTask;
         }
     }
